Logger: remove printout.
The printout present here did not really serve any purpose, and did
not fit with the rest of the class, which did no printouts of its
own accord. Removing it will allow us to, for example, have usage
texts without that annoying "Closing logger" line when DEBUG has
been enabled.